requirements.txt
# AI Framework - Consolidated Requirements
# Production-ready dependencies for the unified system

# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.23.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.13.0
pgvector>=0.2.5
psycopg2-binary>=2.9.9

# HTTP client and utilities
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0

# LLM providers
openai>=1.3.0
google-generativeai>=0.3.0

# Google Cloud libraries for Drive integration
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0

# Document processing
pypdf2>=3.0.0

# Embeddings (local, no API required)
sentence-transformers>=2.2.0

# Basic LangChain for agent framework
langchain-core>=0.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For test client

# Monitoring and metrics
prometheus_client>=0.19.0

# Utilities
python-multipart>=0.0.6  # For file uploads